requirements.txt
fastapi
uvicorn
sqlalchemy
loguru
requests
python-dotenv